SQLAlchemy cleanup to provide better 1.4 compatibility
