+++ tests/integration_tests/test_sqlalchemy/test_reflect.py
@@ -7,8 +7,8 @@ from clickhouse_connect.cc_sqlalchemy.datatypes.sqltypes import UInt32
 
 def test_basic_reflection(test_engine: Engine):
     conn = test_engine.connect()
-    metadata = db.MetaData(bind=test_engine, reflect=True, schema='system')
-    table = db.Table('tables', metadata)
+    metadata = db.MetaData(bind=test_engine, schema='system')
+    table = db.Table('tables', metadata, autoload_with=test_engine)
     query = db.select([table.columns.create_table_query])
     result = conn.execute(query)
     rows = result.fetchmany(100)
@@ -20,8 +20,8 @@ def test_full_table_reflection(test_engine: Engine, test_db: str, test_table_eng
     conn.execute(
         'CREATE TABLE IF NOT EXISTS reflect_test (key UInt32, value FixedString(20))' +
         f'ENGINE {test_table_engine} ORDER BY key')
-    metadata = db.MetaData(bind=test_engine, reflect=True, schema=test_db)
-    table = db.Table('reflect_test', metadata)
+    metadata = db.MetaData(bind=test_engine, schema=test_db)
+    table = db.Table('reflect_test', metadata, autoload_with=test_engine)
     assert table.columns.key.type.__class__ == UInt32
     assert 'MergeTree' in table.engine.name
